NET-5: moved routing.Resolver -> network.RoutingTable
