--- network/routing/resolver.go
@@ -1,37 +1,0 @@
-/*
- *    Copyright 2018 Insolar
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-package routing
-
-import (
-	"github.com/insolar/insolar/core"
-	"github.com/insolar/insolar/network/transport/host"
-)
-
-// PartitionPolicy contains all rules how to initiate globule resharding.
-type PartitionPolicy interface {
-	ShardsCount() int
-}
-
-// Resolver contains all routing information of the network.
-type Resolver interface {
-	// Resolve NodeID -> Address. Can initiate network requests.
-	Resolve(core.RecordRef) (string, error)
-	// AddToKnownHosts add host to routing table.
-	AddToKnownHosts(*host.Host)
-	// Rebalance recreate shards of routing table with known hosts according to new partition policy.
-	Rebalance(PartitionPolicy)
-}
